LU-19079 ssk: use sha256 nodemap name from client
When the client mounts and an srpc flavor is enforced, the client
needs to authenticate. It forges the authentication token by calling
lgss_keyring in userspace (called from request_key(), based on the
lgssc key type).
In the case of Shared Secret Key authentication (SSK), the
authentication token contains several fields (see lgss_sk_using_cred),
including the sha256 of the nodemap name as found in the SSK key
('-n' option to lgss_sk). This authentication token is attached to the
SEC_CTX_INIT RPC sent to the server (MGS, MDS, OSS).
On server side, the authentication token is retrieved from the RPC in
the kernel, and handed over to the userspace daemon lsvcgssd. This
daemon is responsible for checking client authentication. In the case
of SSK, among other aspects (HMAC verification, ...) the daemon
recomputes the sha256 of the nodemap name, and compares it with the
one sent by the client.

If the lsvcgssd daemon fails to verify the sha256 of the nodemap name,
embedded into the authentication request by the client as part of the
'netstring', against the nodemap inferred from the client NID, it now
retrieves the nodemap name by doing a nodemap lookup from the sha256
of the nodemap name sent by the client (nodemap_lookup_by_sha()
issuing the LCFG_NODEMAP_LOOKUP_SHA ioctl). The kernel returns the
corresponding nodemap name, only if it has the gssonly_identification
property. Then we carry out the rest of the authentication process
with this nodemap name. We finally put this nodemap name in the rsc
downcall data, to let the kernel know it should be using it.

This is safe to use the sha sent by the client in the 'netstring', as
it is protected from forgery thanks to the HMAC. If the nodemap name
claimed by the client is faked, the HMAC verification would fail on
server side as it would be carried out with a different SSK key.

Test-Parameters: trivial testgroup=review-dne-selinux-ssk-part-2
Signed-off-by: Sebastien Buisson <sbuisson@ddn.com>
Change-Id: I52e2f32933d5946932244df3d2440db98347fb33
Reviewed-by: Andreas Dilger <adilger@whamcloud.com>
Reviewed-by: Marc Vef <mvef@whamcloud.com>
Reviewed-on: https://review.whamcloud.com/c/fs/lustre-release/+/61740
Reviewed-by: Andreas Dilger <adilger@thelustrecollective.com>
Tested-by: jenkins <devops@whamcloud.com>
Tested-by: Maloo <maloo@whamcloud.com>
Reviewed-by: Oleg Drokin <green@whamcloud.com>