LU-12345 docs: add '%k' format to lfs-find.1 man page
Document the '%k' format for the 'lfs find --printf' option, which
prints the number of 1K blocks allocated for the file.

Also fix a number of style issues and typos in the man page.

Fix the --printf formatting options to be ordered alphabetically
to make the options easier to find.

Test-Parameters: trivial
Fixes: 6b8e97b76c ("LU-10378 utils: add formatted printf to lfs find")
Signed-off-by: Andreas Dilger <adilger@whamcloud.com>
Change-Id: Ied07344294d2f91f457c9cc940c919e9f12540e5
Reviewed-on: https://review.whamcloud.com/c/fs/lustre-release/+/60690
Reviewed-by: Oleg Drokin <green@whamcloud.com>
Reviewed-by: Anjus George <georgea@ornl.gov>
Reviewed-by: Frederick Dilger <fdilger@whamcloud.com>
Reviewed-by: Vitaliy Kuznetsov <vkuznetsov@ddn.com>
Tested-by: jenkins <devops@whamcloud.com>
Tested-by: Maloo <maloo@whamcloud.com>