LU-16138 osd-ldiskfs: remove block integrity kernel patches
Rewrite osd_integrity.c to remove the need for block integrity kernel
patches.

Now we do the equivalent of bio_integrity_prep() ourselves in
osd_bio_integrity_handle(), without setting BIP_BLOCK_INTEGRITY flag
on integrity payload to preserve the metadata buffer.

For a WRITE we fill the metadata buffer with guard tags from
ptlrpc if there's any, then submit the bio.
For a READ we need to take over the data integrity
verification when the I/O is completed, during verification we can reuse
the guard tags for ptlrpc, this is done on a per-OSD workqueue.

With the above changes we can finally remove the custom block integrity
kernel patches, and still be able to reuse the guard tags for ptlrpc.

Tested on rhel7.9, rhel8.10 and rocky9.X

Change-Id: I3df1330e00aca59226763f7c94e16911afe0f5b1
Signed-off-by: Li Dongyang <dongyangli@ddn.com>
Reviewed-on: https://review.whamcloud.com/c/fs/lustre-release/+/59549
Reviewed-by: Andreas Dilger <adilger@whamcloud.com>
Tested-by: Maloo <maloo@whamcloud.com>
Reviewed-by: Li Xi <lixi@ddn.com>
Reviewed-by: Shaun Tancheff <shaun.tancheff@hpe.com>
Tested-by: jenkins <devops@whamcloud.com>
Reviewed-by: Oleg Drokin <green@whamcloud.com>