LU-18242 utils: allow 'lfs df -m/-o' to specify one MDT/OST
Extend the 'lfs df' command to allow specifying a single MDT
or OST index with the -m and -o options. This allows users to
easily check the space usage of a specific target without
having to use grep.

For example:
  lfs df -m0 /mnt/lustre  # Show only MDT0
  lfs df -o1 /mnt/lustre  # Show only OST1

This is useful in test scripts and for end users who want to
check the space usage of a specific target.

Signed-off-by: Chakshu Kansal <ckansal@ddn.com>
Change-Id: I74538b6d8c95c41a7d9030f94103a7c0fd8756db
Reviewed-on: https://review.whamcloud.com/c/fs/lustre-release/+/58729
Tested-by: jenkins <devops@whamcloud.com>
Tested-by: Maloo <maloo@whamcloud.com>
Reviewed-by: Andreas Dilger <adilger@whamcloud.com>
Reviewed-by: Oleg Drokin <green@whamcloud.com>