LU-17231 doc: Create script to generate man pages
Created generate-param.sh script that can be used to generate
the bare bones of a man pages with the appropriate sections included.

In particular this script will be used to create pages for the new
lustre parameter man pages, which there are several hundred.

This script can be used as follows:
  generate-manpage.sh NAME SECTION [OUTPUT_FILE]
or
  generate-manpage.sh [DIR/]NAME.(1-8)

for example:
  generate-manpage.sh lctl-list_param 8
will generate a single man page for the command 'lctl list_param'
under the folder Documentation/man8.

or

  generate-manpage.sh lctl-list_param 8 lustre/doc/lctl-list_param.8
will generate the man page in lustre/doc/lctl-list_param.8 instead
of the default directory.

or

  generate-manpage.sh lctl-list_param.8
will generate the man page in lctl-list_param.8 in the current
directory.

Test-Parameters: trivial
Signed-off-by: Frederick Dilger <fdilger@whamcloud.com>
Change-Id: I9b3ec2bf82aa947038aa9c1fb681f6573becec9d
Reviewed-on: https://review.whamcloud.com/c/fs/lustre-release/+/60047
Tested-by: Maloo <maloo@whamcloud.com>
Tested-by: jenkins <devops@whamcloud.com>
Reviewed-by: Arshad Hussain <arshad.hussain@aeoncomputing.com>
Reviewed-by: Oleg Drokin <green@whamcloud.com>
Reviewed-by: Alex Deiter <adeiter@ddn.com>