LU-19121 ldiskfs: check fast symlink for ea_inode correctly
The check for a fast symlink in the presence of only an external
xattr inode is incorrect. If a fast symlink has no xattr block
(i_file_acl == 0), but has an external xattr inode that increases
inode i_blocks, then the check for a fast symlink will incorrectly
fail and __ext4_iget()->ext4_ind_check_inode() will report an error
when the inode is next loaded into memory:

    # ln -s foo /mnt/tmp/bar
    # setfattr -h -n trusted.test \
               -v "$(head -200 /etc/services)" /mnt/tmp/bar
    # umount /mnt/tmp
    # mount /mnt/tmp
    # ls -l /mnt/tmp
    ls: cannot access '/mnt/tmp/bar': Structure needs cleaning
    total 4
     ? l?????????? ? ?    ?        ?            ? bar
    # dmesg | tail -1
    EXT4-fs error (device dm-8): __ext4_iget:5098:
        inode #24578: block 7303014: comm ls: invalid block

The ext4_inode_is_fast_symlink() check should use the superblock
EA_INODE feature flag, not the inode EXT4_EA_INODE_FL flag, since
the latter is only set on the xattr inode itself, and not on the
inode that references it.

Signed-off-by: Andreas Dilger <adilger@whamcloud.com>
Change-Id: Ia7c1c59fcc005a3be11435325a96b809592540e5
Reviewed-on: https://review.whamcloud.com/c/fs/lustre-release/+/59879
Tested-by: jenkins <devops@whamcloud.com>
Tested-by: Maloo <maloo@whamcloud.com>
Reviewed-by: Li Dongyang <dongyangli@ddn.com>
Reviewed-by: Alex Zhuravlev <bzzz@whamcloud.com>
Reviewed-by: Oleg Drokin <green@whamcloud.com>