LU-17431 nodemap: find innermost nid range
When searching for an exact NID range matching [ start_nid, end_nid ],
we must return first the innermost range associated with a possibly
dynamic nodemap, as it is going to be used preferably for association
of clients with nodemaps.
For that purpose, introduce a new struct nodemap_range_tree *rn_tree
field in struct lu_nid_range, that points to the range tree associated
with the nid range.

Fixes: e9278b8da1 ("LU-17431 nodemap: make dynamic nodemaps hierarchical")
Signed-off-by: Sebastien Buisson <sbuisson@ddn.com>
Change-Id: Ide5d69794adb3516603b6594aad5f8f8ba123cc5
Reviewed-on: https://review.whamcloud.com/c/fs/lustre-release/+/59890
Tested-by: Maloo <maloo@whamcloud.com>
Reviewed-by: Oleg Drokin <green@whamcloud.com>
Reviewed-by: Andreas Dilger <adilger@whamcloud.com>
Reviewed-by: Marc Vef <mvef@whamcloud.com>
Tested-by: jenkins <devops@whamcloud.com>