LU-19126 lnet: Refactor jt_import netlink code
The netlink code in jt_import contains nearly identical code for the
"net", "peer", "route", "discover", and "ping" yaml sequences. This
has been consolidated into a table-driven approach using the new
command_mapping structure.

Code for handling the global parameter is split into a helper function
to make the main while loop more readable.

sanity-lnet test_255 contains a test of the import command, but it
was not written correctly and with the current code it silently fails.
The test has been corrected so that LNet is correctly initialized
prior to executing the import command.

sanity-lnet test_350 is added to exercise the --exec import option
using the discover and ping commands.

Test-Parameters: trivial
Signed-off-by: Chris Horn <chris.horn@hpe.com>
Change-Id: I26ecb465c22cbab871532ba8966761cc965b6c2b
Reviewed-on: https://review.whamcloud.com/c/fs/lustre-release/+/59928
Reviewed-by: James Simmons <jsimmons@infradead.org>
Tested-by: jenkins <devops@whamcloud.com>
Reviewed-by: Oleg Drokin <green@whamcloud.com>
Reviewed-by: Andreas Dilger <adilger@whamcloud.com>
Tested-by: Maloo <maloo@whamcloud.com>
Tested-by: James Simmons <jsimmons@infradead.org>