LU-19129 utils: message for lctl-nodemap on non-MGS node
In case an lctl-nodemap command fails because it was launched on the
wrong node, the error message should explicitly inform the user that
the command must be executed on the MGS node.
Make nodemap handling code in the kernel return -ENXIO in this case,
so that userspace can print the appropriate message.

Test-Parameters: trivial
Signed-off-by: Sebastien Buisson <sbuisson@ddn.com>
Change-Id: Ieda9dd2dbd2df3f6b1be3026113bf75867483cbd
Reviewed-on: https://review.whamcloud.com/c/fs/lustre-release/+/59932
Reviewed-by: Andreas Dilger <adilger@whamcloud.com>
Reviewed-by: Marc Vef <mvef@whamcloud.com>
Tested-by: jenkins <devops@whamcloud.com>
Tested-by: Maloo <maloo@whamcloud.com>
Reviewed-by: Oleg Drokin <green@whamcloud.com>