LU-17431 nodemap: Fileset modification on dynamic nodemaps
This patch allows "lctl nodemap_fileset_modify" to be used on dynamic
nodemaps. In-place fileset updates are important for dynamic nodemaps
because, depending on the parent nodemap's filesets, deleting and
re-adding a fileset on the dynamic nodemap is not always possible. As
a result, the deprecated "lctl nodemap_set_fileset" command must be
used previously for primary filesets which does not allow changing the
read-only flag. Singular, alternate filesets could not be modified on
the child nodemap.

"lctl nodemap_fileset_modify" allows child nodemaps to freely
configure their filesets independent of the parent as long as the
child operates within the namespace restrictions of the parent's
filesets. This includes that parent's read-only filesets cannot become
read-write in a child nodemap.

Sanity-sec test 72f was added to test this feature.

Test-Parameters: trivial testlist=sanity-sec
Signed-off-by: Marc Vef <mvef@whamcloud.com>
Change-Id: Ib132a9e5bb3d1f06369ebdda8f8500c7317bc245
Reviewed-on: https://review.whamcloud.com/c/fs/lustre-release/+/59999
Tested-by: jenkins <devops@whamcloud.com>
Reviewed-by: Sebastien Buisson <sbuisson@ddn.com>
Tested-by: Maloo <maloo@whamcloud.com>
Reviewed-by: Andreas Dilger <adilger@whamcloud.com>
Reviewed-by: Oleg Drokin <green@whamcloud.com>