LU-19157 nodemap: introduce range type
Introduce a type for nodemap ranges. The regular type is for usual
nodemap ranges, and a new type is used for nodemap ranges that define
NID ban lists.
NM_RANGE_FL_REG = 0x0
NM_RANGE_FL_BAN = 0x1
The first 4 bits of nk_range_id are now reserved for the range type.
This is fine for upgraded systems that are not aware of the nodemap
range type, as the value for the regular nodemap range type is 0.
And this leaves us with 28 bits to store the range id, for a max range
id of 268435455.

Test-Parameters: trivial
Signed-off-by: Sebastien Buisson <sbuisson@ddn.com>
Change-Id: Ibf019059b0d7f442b514c38b8890a4ba9764b3da
Reviewed-on: https://review.whamcloud.com/c/fs/lustre-release/+/60027
Tested-by: jenkins <devops@whamcloud.com>
Reviewed-by: Marc Vef <mvef@whamcloud.com>
Reviewed-by: Andreas Dilger <adilger@whamcloud.com>
Reviewed-by: Oleg Drokin <green@whamcloud.com>
Tested-by: Maloo <maloo@whamcloud.com>