LU-19157 nodemap: add banlist property
Add a new 'banlist' property to nodemaps. Its syntax is similar to the
'ranges' property, and can be managed from userspace thanks to new
commands:
lctl nodemap_banlist_add --name NODEMAP_NAME --range NID_RANGE
lctl nodemap_banlist_del --name NODEMAP_NAME --range NID_RANGE
NID ranges for the 'banlist' of a nodemap have to be included in the
'ranges' of the nodemap: only clients that belong to this nodemap can
be added to the ban list.

Test-Parameters: trivial
Signed-off-by: Sebastien Buisson <sbuisson@ddn.com>
Change-Id: I60061420ee6f3de305b5b54b6a04df3dcbbd2d0a
Reviewed-on: https://review.whamcloud.com/c/fs/lustre-release/+/60028
Reviewed-by: Marc Vef <mvef@whamcloud.com>
Tested-by: jenkins <devops@whamcloud.com>
Reviewed-by: Oleg Drokin <green@whamcloud.com>
Reviewed-by: Andreas Dilger <adilger@whamcloud.com>
Tested-by: Maloo <maloo@whamcloud.com>