LU-17431 nodemap: do not revoke global lock for dyn nodemap
In case of an MGS/MDS combo, or when the MGS and the MDS run on the
same node, we need to detect if the 'lctl nodemap_xxx' operation is
carried out on a regular or dynamic nodemap. In case of a dynamic
nodemap, we should not revoke the global lock on the nodemap
configuration, otherwise it would lead to the loss of any existing
dynamic nodemaps on all servers.

Signed-off-by: Sebastien Buisson <sbuisson@ddn.com>
Change-Id: Ia56ba535eecda95dc0968a133c6d732a4847e24d
Reviewed-on: https://review.whamcloud.com/c/fs/lustre-release/+/60044
Tested-by: jenkins <devops@whamcloud.com>
Tested-by: Maloo <maloo@whamcloud.com>
Reviewed-by: Oleg Drokin <green@whamcloud.com>
Reviewed-by: Andreas Dilger <adilger@whamcloud.com>
Reviewed-by: Marc Vef <mvef@whamcloud.com>