LU-19157 nodemap: allow ban list on default nodemap
Allow the default nodemap to have a ban list. As the default nodemap
does not have any NID ranges, the constraint on its ban list is that
it cannot conflict with any NID range defined on any other nodemap.

Enhance sanity-sec test_81a to test ban list on the default nodemap.

Signed-off-by: Sebastien Buisson <sbuisson@ddn.com>
Change-Id: I4c3d6f86353aa5f4ec00750891d9b47511b4e384
Reviewed-on: https://review.whamcloud.com/c/fs/lustre-release/+/60074
Reviewed-by: Andreas Dilger <adilger@whamcloud.com>
Tested-by: jenkins <devops@whamcloud.com>
Reviewed-by: Oleg Drokin <green@whamcloud.com>
Reviewed-by: Marc Vef <mvef@whamcloud.com>
Tested-by: Maloo <maloo@whamcloud.com>