LU-18982 lov: Do not display lmm_stripe_offset for mdt file
The command lfs getstripe <file> displays lmm_stripe_offset
as 0 for DOM files, which can be confusing.

The lmm_stripe_offset points to the OST index where the file's
data resides. However, for DOM files, the first component is
stored on the MDT, not an OST. This discrepancy in
interpretation can lead to understanding that DOM resides
on OST 0, and in case DOM resides on MDT indext 1 it causes
user to think DOM is on MDT 0.

Added test_909() to test the change

Test-Parameters: trivial
HPE-bug-id:LUS-12836
Signed-off-by: Rajeev Mishra <rajeevm@hpe.com>
Change-Id: I6905c3eeaba258e87b68aaab28ca7d2f294a7044
Reviewed-on: https://review.whamcloud.com/c/fs/lustre-release/+/59252
Reviewed-by: Shaun Tancheff <shaun.tancheff@hpe.com>
Reviewed-by: Oleg Drokin <green@whamcloud.com>
Reviewed-by: Andreas Dilger <adilger@whamcloud.com>
Reviewed-by: Petros Koutoupis <petros.koutoupis@hpe.com>
Tested-by: jenkins <devops@whamcloud.com>
Tested-by: Maloo <maloo@whamcloud.com>