LU-19157 nodemap: forward ban status to client
Use new pb_flags value MSG_CLIENT_BANNED to foward ban status to the
client in the repmsg buffer. The client is not nodemap aware, but can
print a console warning that it was banned by the server. This message
is limited to one occurence every 6 hours to avoid flooding.

sanity-sec test_81a is enhanced to check the message is printed on the
client console.

Signed-off-by: Sebastien Buisson <sbuisson@ddn.com>
Change-Id: I4b2e407b8c91d0af723896cd2e6785e0b6d5b1ef
Reviewed-on: https://review.whamcloud.com/c/fs/lustre-release/+/60122
Reviewed-by: Andreas Dilger <adilger@whamcloud.com>
Reviewed-by: Marc Vef <mvef@whamcloud.com>
Tested-by: Maloo <maloo@whamcloud.com>
Tested-by: jenkins <devops@whamcloud.com>
Reviewed-by: Oleg Drokin <green@whamcloud.com>