LU-17716 utils: add "tunefs.lustre --replace" option
If a new target is not formatted with "mkfs.lustre --replace"
for some reason, then it would be useful to have a
"tunefs.lustre --replace" option to reset the OST config,
label, etc. so that it can be re-added to the filesystem
without trying to register with the MGS again.

This patch makes the "--replace" option work for both
mkfs.lustre and tunefs.lustre. With this option specified,
the LDD_F_VIRGIN (never registered) flag will be cleared,
then the newly-formatted OST will be able to use the same
index as the previously used target and skip MGS registration.

Test-Parameters: trivial fstype=ldiskfs env=SLOW=yes,ONLY=69 testlist=conf-sanity
Test-Parameters: trivial fstype=zfs env=SLOW=yes,ONLY=69 testlist=conf-sanity

Change-Id: I1a344dda2f448bd38c22ed202767eeb7fcd97edf
Signed-off-by: Jian Yu <yujian@whamcloud.com>
Reviewed-on: https://review.whamcloud.com/c/fs/lustre-release/+/59035
Tested-by: Maloo <maloo@whamcloud.com>
Reviewed-by: Oleg Drokin <green@whamcloud.com>
Tested-by: jenkins <devops@whamcloud.com>
Reviewed-by: Andreas Dilger <adilger@whamcloud.com>
Reviewed-by: Mikhail Pershin <mpershin@whamcloud.com>