LU-17980 ldiskfs: add fstrim mount option
Introduce ldiskfs fstrim mount option. to improve online discard
behavior.
-o fstrim overrides -o discard, instead of doing online discard,
when a block group has more free blocks than the tunable
bg_trimmed_threshold, it will be scheduled for discard for the
whole group with a delay, using ext4_trim_all_free().

When more blocks are freed in the block group, the delay will be renewed,
to aggregate the potential future free blocks, but not longer
than the tunable 60 seconds from the original discard request.

When a block group is scheduled for discard, it will be skipped
in the block allocator for the cr0 and cr1 scan, to avoid repeat
free/allocate on the block group. If we can't find any suitable
block group in the cr0 and cr1 scan, we will try cr0 and cr1 again
but this time do not skip the block groups.

-o fstrim also accepts an optional fstrim_min_blocks parameter
which is also tunable via sysfs, to control the min free blocks
we trim in ext4_trim_all_free().
If the group's average free fragments is less than fstrim_min_blocks,
we skip the group as it's too fragmented and will result in many
small trim commands to the storage.

There's another sysfs tunable fstrim_avg_frag_no_delay,
if the group's average free fragments is greater than
fstrim_avg_frag_no_delay, we try to trim the group right away without
delaying and gathering more free blocks.

Fix lmd_parse() to pass down the ldiskfs mount options to OST,
when there's mgsnode option the current code cut off options string,
losing the extra ldiskfs mount options.

Change-Id: I7f9bc0c10f5caf3ded72b781684579fdd5d796fb
Signed-off-by: Li Dongyang <dongyangli@ddn.com>
Reviewed-on: https://review.whamcloud.com/c/fs/lustre-release/+/60235
Reviewed-by: Oleg Drokin <green@whamcloud.com>
Reviewed-by: Alex Zhuravlev <bzzz@whamcloud.com>
Tested-by: Eugene Birkine <ebirkine@ddn.com>
Tested-by: Maloo <maloo@whamcloud.com>
Reviewed-by: Jason Huselton <jhuselton@ddn.com>
Tested-by: jenkins <devops@whamcloud.com>
Reviewed-by: Andreas Dilger <adilger@thelustrecollective.com>