LU-18756 ofd: add repair resource id implementation
This patch adds the implementation for the background repair thread to
modify the OST object ID attributes based on the passed "lu_fid".
Before the thread updates the IDs, it ensures that IDs, that were
written in the meantime, are not overwritten.

The introduced "ofd_repair_resource_ids()" function uses the incoming
client RPC IDs to set them on the OST objects. Here, the function
relies on "struct obdo" to use the client's IDs that are set as valid
by the "o_valid" field. Further, it is possible to repair only a
subset of IDs depending on which object's IDs are set and which obdo
IDs are available.

Because of the background repair thread, ID repair can be included in
read-only operations, if the OST itself is not mounted as read-only.

Test-Parameters: trivial
Signed-off-by: Marc Vef <mvef@whamcloud.com>
Change-Id: Iae7dd17457608143bdb3aa023429a4ba4586a699
Reviewed-on: https://review.whamcloud.com/c/fs/lustre-release/+/60301
Reviewed-by: Andreas Dilger <adilger@whamcloud.com>
Tested-by: jenkins <devops@whamcloud.com>
Tested-by: Maloo <maloo@whamcloud.com>
Reviewed-by: Sebastien Buisson <sbuisson@ddn.com>
Reviewed-by: Alex Zhuravlev <bzzz@whamcloud.com>