LU-18756 ofd: include resource id repair into id check
This patch reorganizes the resource id check on the OST and to include
the id repair into the id check, allowing the repair to be triggered
when the id check returns -EAGAIN, indicating that the OST object has
unset resource ids.

In all cases, the object's attributes are only retrieved once for the
ID check and/or ID repair.

Further, "struct ofd_object" is now explicitly passed to id check and
repair functions instead of relying on "ofd_info(env)->fti_obj".

This patch does not include the ID repair into the code path yet.

Test-Parameters: trivial
Test-Parameters: testlist=sanity-sec env=ONLY="75"
Signed-off-by: Marc Vef <mvef@whamcloud.com>
Change-Id: Ieb805ec6f2f3fc64c5292cd48a62c0a1c298c284
Reviewed-on: https://review.whamcloud.com/c/fs/lustre-release/+/60303
Reviewed-by: Oleg Drokin <green@whamcloud.com>
Tested-by: jenkins <devops@whamcloud.com>
Tested-by: Maloo <maloo@whamcloud.com>
Reviewed-by: Sebastien Buisson <sbuisson@ddn.com>
Reviewed-by: Andreas Dilger <adilger@whamcloud.com>