LU-11432 lnet: Assign specific values to enum constants
The enum constants in lnet-types.h are automatically
enumerated from 0. This can break the network
protocol compatibility if mistakenly a new
values is inserted or a value is deleted in the
middle of the list.

This patch also adds a new enum struct lnet_md_options
to pack all the lnet MD options together.

Test-Parameters: trivial
Change-Id: If3c2155447056d09b60eb7d4a4b992b4ee8b2787
Signed-off-by: Sonia Sharma <sharmaso@whamcloud.com>
Signed-off-by: Aryan Gupta <argupta@ddn.com>
Reviewed-on: https://review.whamcloud.com/c/fs/lustre-release/+/33251
Reviewed-by: James Simmons <jsimmons@infradead.org>
Reviewed-by: Oleg Drokin <green@whamcloud.com>
Reviewed-by: Shaun Tancheff <shaun.tancheff@hpe.com>
Reviewed-by: Andreas Dilger <adilger@whamcloud.com>
Tested-by: jenkins <devops@whamcloud.com>
Tested-by: Maloo <maloo@whamcloud.com>