LU-19209 utils: ofd_access_log_reader usage mention log_size
ofd access logs are not enabled by default; they are enabled by
setting the size of the circular buffer they require.

State this in the usage message of the utility used to read them, and
in the comment of the source file implementing them, so users need not
spelunk through code to learn how.

Test-Parameters: trivial
Signed-off-by: Olaf Faaland <faaland1@llnl.gov>
Change-Id: I048e53fdf5c91d3302baaec9b0cad275269aa168
Reviewed-on: https://review.whamcloud.com/c/fs/lustre-release/+/60356
Tested-by: jenkins <devops@whamcloud.com>
Tested-by: Maloo <maloo@whamcloud.com>
Reviewed-by: Oleg Drokin <green@whamcloud.com>
Reviewed-by: Andreas Dilger <adilger@whamcloud.com>
Reviewed-by: Alex Zhuravlev <bzzz@whamcloud.com>
Reviewed-by: Frederick Dilger <fdilger@whamcloud.com>