LU-17814 utils: Enable pfind by default
Add error handling and enable pfind by default.

This requires switching find_skip_file from counting to
probability based, otherwise we'd need an expensive atomic
op to coordinate the exact skip count.  But the point of the
option is statistical rather than exact behavior.

Increase the number of files in the relevant test to get
the law of large numbers in our favor.  Some complex math
suggests the probability of failure is extremely small.

Signed-off-by: Patrick Farrell <pfarrell@whamcloud.com>
Change-Id: I3dcbe94483873bd84d16dc02a860d70d314b9964
Reviewed-on: https://review.whamcloud.com/c/fs/lustre-release/+/57306
Reviewed-by: Andreas Dilger <adilger@whamcloud.com>
Reviewed-by: Marc Vef <mvef@whamcloud.com>
Tested-by: Maloo <maloo@whamcloud.com>
Tested-by: jenkins <devops@whamcloud.com>
Reviewed-by: Oleg Drokin <green@whamcloud.com>